+++ client/src/components/Banners/Release/ReleaseField.vue
@@ -40,7 +40,7 @@ export default {
     isTimeBased: { type: Boolean, default: true },
     time: { type: String, default: '' },
     isCloseable: { type: Boolean, default: true },
-    links: { type: Array, default: () => [{ urlShortHand: 'releaseNotes', text: 'menu.entry.release-notes' }] },
+    links: { type: Array, default: () => [{ urlShortHand: 'release_notes', text: 'menu.entry.release-notes' }] },
   },
   data () {
     return {}