+++ client/src/components/Topbar/Items/User/MenuUser.vue
@@ -23,7 +23,7 @@
     >
       <a
         v-if="isBeta"
-        :href="$url('releaseNotes')"
+        :href="$url('release_notes')"
         role="menuitem"
         class="dropdown-item dropdown-action list-group-item-warning"
       >