+++ client/src/views/partials/Footer/Footer.vue
@@ -336,7 +336,7 @@
           </a>
           <a
             v-else
-            :href="$url('releaseNotes')"
+            :href="$url('release_notes')"
             v-html="$i18n('releases.2022-05')"
           />
           <a
@@ -415,7 +415,7 @@ export default {
           name: 'menu.entry.aboutUs',
           items: [
             { url: 'mission', title: 'menu.entry.mission' },
-            { url: 'grundsaetze', title: 'menu.entry.fundamentals' },
+            { url: 'wiki_grundsaetze', title: 'menu.entry.fundamentals' },
             { url: 'blog', title: 'menu.entry.blog' },
             { url: 'team', title: 'menu.entry.team' },
             { url: 'partner', title: 'menu.entry.partners' },
@@ -425,11 +425,11 @@ export default {
         {
           name: 'menu.entry.background',
           items: [
-            { url: 'support', title: 'menu.entry.support', target: '_blank' },
+            { url: 'freshdesk', title: 'menu.entry.support', target: '_blank' },
             { url: 'wiki', title: 'menu.entry.wiki' },
-            { url: 'guide', title: 'menu.entry.guide', target: '_blank' },
+            { url: 'wiki_guide', title: 'menu.entry.guide', target: '_blank' },
             { url: 'statistics', title: 'menu.entry.statistics' },
-            { url: 'releaseNotes', title: 'menu.entry.release-notes' },
+            { url: 'release_notes', title: 'menu.entry.release-notes' },
           ],
         },
         {