+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 - Rebuild store options to vue #2231 @chriswalg
 - Changed deleting wall posts to the API #798 !2253 @alex.simm
 - Rebuild profile infos to vue !2276 !2287 @chriswalg
+- Rebuild join region to vue !2294 @sefn
 
 ## Dev/Test/CI stuff
 - bump node to version 16 !2135 @peter.toennies